Refactor for production readiness and fix identified issues
- Consolidated redundant user routers in API v1.
- Secured database initialization to prevent accidental data loss in production.
- Added Dockerfile and docker-compose.yml for containerization.
- Added .env.example for secure configuration management.
- Implemented a global exception handler for consistent production error reporting.
- Fixed inconsistent imports and minor bugs in services.
- Added a comprehensive production deployment guide in docs/PRODUCTION.md.
