+++ backend/.env.example
@@ -1,0 +1,28 @@
+# Project Settings
+PROJECT_NAME="eCyber"
+VERSION="1.0.0"
+DEBUG=False
+PRODUCTION=True
+DOCS=False
+
+# Database
+# For SQLite (Default):
+SQLALCHEMY_DATABASE_URL="sqlite+aiosqlite:///./security.db"
+# For PostgreSQL:
+# SQLALCHEMY_DATABASE_URL="postgresql+asyncpg://user:password@localhost/dbname"
+
+# Security
+SECRET_KEY="your-secret-key-for-jwt-tokens"
+ACCESS_TOKEN_EXPIRE_MINUTES=10080  # 7 days
+
+# CORS
+# BACKEND_CORS_ORIGINS='["http://localhost:3000", "http://localhost:4000"]'
+
+# Network
+NETWORK_INTERFACE="Wi-Fi"
+
+# Redis (Optional, for production scaling)
+# REDIS_URL="redis://localhost:6379/0"
+
+# GeoIP (Optional, requires MaxMind database)
+# GEOIP_DATABASE_PATH="data/GeoLite2-Country.mmdb"
