+++ backend/app/core/config.py
@@ -14,7 +14,7 @@ class Settings(BaseSettings):
     BACKEND_CORS_ORIGINS: List[str] = [
         "http://localhost:3000",
         "http://127.0.0.1:3000",
-        "http://localhost:4000 ",
+        "http://localhost:4000",
         "http://127.0.0.1:4000",
     ]
