+++ backend/requirements.txt
@@ -130,7 +130,7 @@ sympy==1.14.0
 tensorboard==2.19.0
 tensorboard-data-server==0.7.2
 tensorflow==2.19.0
-tensorflow-io-gcs-filesystem==0.31.0
+tensorflow-io-gcs-filesystem
 termcolor==3.1.0
 threadpoolctl==3.6.0
 tldextract==5.3.0
